+++ changelog.md
@@ -1,2 +1,2 @@
-* AnimationManager now respect HOLD_ON_LAST_FRAME animation type. Again
-* Fixed problem with objects with non zero base rotation angle
+* Added first version of render layers
+* Added some small optimization